Let's sleep :))

+++ PlayerGUITabs.cpp
@@ -39,7 +39,7 @@ const bool PlayerGUITabs::tabsOpen()
 {
 	//Loop through tabs and check if any are open
 	bool open = false;
-	for (size_t i = 0; i < this->tabs.size() && !open; i++)
+	for (int i = 0; i < this->tabs.size() && !open; i++)
 	{
 		if (this->tabs[i]->getOpen())
 			open = true;
@@ -50,22 +50,22 @@ const bool PlayerGUITabs::tabsOpen()
 
 void PlayerGUITabs::toggleTab(const int tab_index)
 {
-	if(tab_index >= 0 || tab_index < this->tabs.size())
+	if (tab_index >= 0 || tab_index < this->tabs.size())
 		this->tabs[tab_index]->toggle();
 }
 
 void PlayerGUITabs::update()
 {
-	for (size_t i = 0; i < this->tabs.size(); i++)
+	for (int i = 0; i < this->tabs.size(); i++)
 	{
-		if(this->tabs[i]->getOpen())
+		if (this->tabs[i]->getOpen())
 			this->tabs[i]->update();
 	}
 }
 
 void PlayerGUITabs::render(sf::RenderTarget & target)
 {
-	for (size_t i = 0; i < this->tabs.size(); i++)
+	for (int i = 0; i < this->tabs.size(); i++)
 	{
 		if (this->tabs[i]->getOpen())
 			this->tabs[i]->render(target);
